# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: resumes
    description: Resume upload and analysis
  - name: jobs
    description: Job role recommendations
  - name: stats
    description: Dashboard statistics
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /resumes:
    get:
      operationId: listResumes
      tags: [resumes]
      summary: List the authenticated user's resumes
      responses:
        "200":
          description: A list of resumes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResumeSummary"
    post:
      operationId: uploadResume
      tags: [resumes]
      summary: Upload a resume PDF as a base64 string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadResumePayload"
      responses:
        "201":
          description: The freshly created resume record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"

  /resumes/{id}:
    get:
      operationId: getResume
      tags: [resumes]
      summary: Get a single resume with full analysis
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: A resume with its analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"
    delete:
      operationId: deleteResume
      tags: [resumes]
      summary: Delete a resume
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Resume deleted

  /resumes/{id}/analyze:
    post:
      operationId: analyzeResume
      tags: [resumes]
      summary: Run AI analysis on a resume (ATS score, missing keywords, improvements, role recommendations)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeResumePayload"
      responses:
        "200":
          description: The updated resume with analysis populated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"

  /jobs/roles:
    get:
      operationId: listJobRoles
      tags: [jobs]
      summary: List the predefined job roles available for matching
      responses:
        "200":
          description: A list of job role definitions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobRole"

  /stats/summary:
    get:
      operationId: getDashboardSummary
      tags: [stats]
      summary: Aggregated counts and averages for the user's dashboard
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    UploadResumePayload:
      type: object
      properties:
        fileName:
          type: string
        fileBase64:
          type: string
          description: Base64-encoded PDF file contents (data URL prefix is stripped server-side)
      required:
        - fileName
        - fileBase64

    AnalyzeResumePayload:
      type: object
      properties:
        targetRole:
          type: ["string", "null"]
          description: Optional target role to bias the analysis toward (e.g. Backend Engineer)

    ResumeSummary:
      type: object
      properties:
        id:
          type: integer
        fileName:
          type: string
        atsScore:
          type: ["integer", "null"]
        topRole:
          type: ["string", "null"]
        status:
          type: string
          description: One of "uploaded" | "analyzing" | "analyzed" | "failed"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - fileName
        - status
        - createdAt

    Resume:
      type: object
      properties:
        id:
          type: integer
        fileName:
          type: string
        rawText:
          type: string
        status:
          type: string
        atsScore:
          type: ["integer", "null"]
        keywordScore:
          type: ["integer", "null"]
        skillScore:
          type: ["integer", "null"]
        experienceScore:
          type: ["integer", "null"]
        extractedSkills:
          type: array
          items:
            type: string
        missingKeywords:
          type: array
          items:
            type: string
        weakAreas:
          type: array
          items:
            type: string
        improvements:
          type: array
          items:
            type: string
        improvedBullets:
          type: array
          items:
            type: string
        recommendedRoles:
          type: array
          items:
            $ref: "#/components/schemas/RecommendedRole"
        summary:
          type: ["string", "null"]
          description: A short AI-written narrative summary of the candidate
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - fileName
        - rawText
        - status
        - extractedSkills
        - missingKeywords
        - weakAreas
        - improvements
        - improvedBullets
        - recommendedRoles
        - createdAt
        - updatedAt

    RecommendedRole:
      type: object
      properties:
        role:
          type: string
        matchPercent:
          type: integer
        missingSkills:
          type: array
          items:
            type: string
        matchedSkills:
          type: array
          items:
            type: string
        roadmap:
          type: array
          items:
            $ref: "#/components/schemas/RoadmapWeek"
      required:
        - role
        - matchPercent
        - missingSkills
        - matchedSkills
        - roadmap

    RoadmapWeek:
      type: object
      properties:
        week:
          type: integer
        focus:
          type: string
        tasks:
          type: array
          items:
            type: string
      required:
        - week
        - focus
        - tasks

    JobRole:
      type: object
      properties:
        slug:
          type: string
        title:
          type: string
        category:
          type: string
        description:
          type: string
        coreSkills:
          type: array
          items:
            type: string
        niceToHaveSkills:
          type: array
          items:
            type: string
      required:
        - slug
        - title
        - category
        - description
        - coreSkills
        - niceToHaveSkills

    DashboardSummary:
      type: object
      properties:
        totalResumes:
          type: integer
        analyzedResumes:
          type: integer
        averageAtsScore:
          type: ["number", "null"]
        bestAtsScore:
          type: ["integer", "null"]
        topRoleMatches:
          type: array
          items:
            $ref: "#/components/schemas/TopRoleMatch"
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/ActivityItem"
      required:
        - totalResumes
        - analyzedResumes
        - topRoleMatches
        - recentActivity

    TopRoleMatch:
      type: object
      properties:
        role:
          type: string
        averageMatch:
          type: number
        appearances:
          type: integer
      required:
        - role
        - averageMatch
        - appearances

    ActivityItem:
      type: object
      properties:
        id:
          type: integer
        fileName:
          type: string
        atsScore:
          type: ["integer", "null"]
        status:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - fileName
        - status
        - createdAt