# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityItem {
  id: number;
  fileName: string;
  /** @nullable */
  atsScore?: number | null;
  status: string;
  createdAt: Date;
}