# lib/api-zod/src/generated/types/analyzeResumePayload.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AnalyzeResumePayload {
  /**
   * Optional target role to bias the analysis toward (e.g. Backend Engineer)
   * @nullable
   */
  targetRole?: string | null;
}