# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItem } from "./activityItem";
import type { TopRoleMatch } from "./topRoleMatch";

export interface DashboardSummary {
  totalResumes: number;
  analyzedResumes: number;
  /** @nullable */
  averageAtsScore?: number | null;
  /** @nullable */
  bestAtsScore?: number | null;
  topRoleMatches: TopRoleMatch[];
  recentActivity: ActivityItem[];
}