# lib/api-zod/src/generated/types/jobRole.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface JobRole {
  slug: string;
  title: string;
  category: string;
  description: string;
  coreSkills: string[];
  niceToHaveSkills: string[];
}