# lib/api-zod/src/generated/types/recommendedRole.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RoadmapWeek } from "./roadmapWeek";

export interface RecommendedRole {
  role: string;
  matchPercent: number;
  missingSkills: string[];
  matchedSkills: string[];
  roadmap: RoadmapWeek[];
}