# lib/api-zod/src/generated/types/resumeSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ResumeSummary {
  id: number;
  fileName: string;
  /** @nullable */
  atsScore?: number | null;
  /** @nullable */
  topRole?: string | null;
  /** One of "uploaded" | "analyzing" | "analyzed" | "failed" */
  status: string;
  createdAt: Date;
}