# lib/api-zod/src/generated/types/uploadResumePayload.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UploadResumePayload {
  fileName: string;
  /** Base64-encoded PDF file contents (data URL prefix is stripped server-side) */
  fileBase64: string;
}